Adds new path for search; adds fallback while searching for term
